gennum함수 go 파일로 분리
